+++ src/rx-database.ts
@@ -26,7 +26,9 @@ import {
     pluginMissing,
     flatClone,
     PROMISE_RESOLVE_FALSE,
-    randomCouchString
+    randomCouchString,
+    ensureNotFalsy,
+    PROMISE_RESOLVE_VOID
 } from './util';
 import {
     newRxError
@@ -126,7 +128,7 @@ export class RxDatabaseBase<
     private subject: Subject<RxChangeEvent> = new Subject();
     private observable$: Observable<RxChangeEvent> = this.subject.asObservable();
     public storageToken?: string;
-    public broadcastChannel$?: Subject<RxChangeEvent>;
+    public broadcastChannel$: Subject<RxChangeEvent> = new Subject();
 
     /**
      * removes all internal collection-info
@@ -196,14 +198,16 @@ export class RxDatabaseBase<
         [key in keyof CreatedCollections]: RxCollectionCreator
     }): Promise<{ [key in keyof CreatedCollections]: RxCollection }> {
         // get local management docs in bulk request
-        const collectionDocs = await this.internalStore.findDocumentsById(
-            Object
-                .keys(collectionCreators)
-                .map(name => {
-                    const schema: RxJsonSchema<any> = (collectionCreators as any)[name].schema;
-                    return _collectionNamePrimary(name, schema);
-                }),
-            false
+        const collectionDocs = await this.lockedRun(
+            () => this.internalStore.findDocumentsById(
+                Object
+                    .keys(collectionCreators)
+                    .map(name => {
+                        const schema: RxJsonSchema<any> = (collectionCreators as any)[name].schema;
+                        return _collectionNamePrimary(name, schema);
+                    }),
+                false
+            )
         );
 
         const internalDocByCollectionName: any = {};
@@ -269,10 +273,11 @@ export class RxDatabaseBase<
             ret[name] = collection;
 
             // add to bulk-docs list
-            if (!internalDocByCollectionName[name]) {
+            const collectionName = _collectionNamePrimary(name as any, collectionCreators[name].schema);
+            if (!internalDocByCollectionName[collectionName]) {
                 bulkPutDocs.push({
                     document: {
-                        collectionName: _collectionNamePrimary(name as any, collectionCreators[name].schema),
+                        collectionName,
                         schemaHash: schemaHashByName[name],
                         schema: collection.schema.normalized,
                         version: collection.schema.version,
@@ -290,9 +295,11 @@ export class RxDatabaseBase<
             }
         });
 
-        // make a single call to the pouchdb instance
+        // make a single write call to the storage instance
         if (bulkPutDocs.length > 0) {
-            await this.internalStore.bulkWrite(bulkPutDocs);
+            await this.lockedRun(
+                () => this.internalStore.bulkWrite(bulkPutDocs)
+            );
         }
 
         return ret;
@@ -302,12 +309,14 @@ export class RxDatabaseBase<
      * delete all data of the collection and its previous versions
      */
     removeCollection(collectionName: string): Promise<void> {
+        let destroyPromise = PROMISE_RESOLVE_VOID;
         if ((this.collections as any)[collectionName]) {
-            (this.collections as any)[collectionName].destroy();
+            destroyPromise = (this.collections as any)[collectionName].destroy();
         }
 
         // remove schemas from internal db
-        return _removeAllOfCollection(this as any, collectionName)
+        return destroyPromise
+            .then(() => _removeAllOfCollection(this as any, collectionName))
             // get all relevant pouchdb-instances
             .then(knownVersions => {
                 return Promise.all(
@@ -572,9 +581,8 @@ export async function _removeAllOfCollection(
 
 function _prepareBroadcastChannel<Collections>(rxDatabase: RxDatabase<Collections>): void {
     if (!rxDatabase.broadcastChannel) {
-        return;
+        throw newRxError('SNH', { args: { rxDatabase } });
     }
-    rxDatabase.broadcastChannel$ = new Subject();
 
     rxDatabase.broadcastChannel.addEventListener('message', (msg: RxChangeEventBroadcastChannelData) => {
         if (msg.storageToken !== rxDatabase.storageToken) {
@@ -587,7 +595,7 @@ function _prepareBroadcastChannel<Collections>(rxDatabase: RxDatabase<Collection
         }
         const changeEvent = msg.cE;
 
-        (rxDatabase.broadcastChannel$ as any).next(changeEvent);
+        rxDatabase.broadcastChannel$.next(changeEvent);
     });
 
     rxDatabase._subs.push(